C++ Magic Square
C++ program that creates an N*N magic square, i.e., a square
arrangement of the numbers 1, 2, , , in which the sum of rows, columns,
and diagonals are the same. The user (i.e., TAs) will specify the size
of the square matrix: N. The value N must be an odd number between 3
and 15.